CMakeLists.txt: Removed libgcc dependencies from DEB packages
The libgcc dependencies were because of <stdbool.h>, but these were outdated
and rather difficult to maintain, as newer versions should be added for every
new version of GCC, and different versions should be added for clang as well.
It is safer to just assume that whosoever uses these headers will have a decent
compiler with <stdbool.h> built in, hence it is safe to drop these dependencies.